requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   openai
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==25.3.0
    # via aiohttp
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
exceptiongroup==1.3.0
    # via anyio
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   ollama
    #   openai
httpx-sse==0.4.1
    # via langchain-community
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.10.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.27
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.3.27
    # via -r requirements.in
langchain-core==0.3.74
    # via
    #   langchain
    #   langchain-community
    #   langchain-ollama
    #   langchain-openai
    #   langchain-sandbox
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-ollama==0.3.6
    # via -r requirements.in
langchain-openai==0.3.30
    # via -r requirements.in
langchain-sandbox==0.0.6
    # via -r requirements.in
langchain-text-splitters==0.3.9
    # via langchain
langgraph==0.6.5
    # via
    #   -r requirements.in
    #   langchain-sandbox
langgraph-checkpoint==2.1.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==0.6.4
    # via langgraph
langgraph-sdk==0.2.0
    # via langgraph
langsmith==0.4.14
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.2.6
    # via
    #   langchain-community
    #   pandas
ollama==0.5.3
    # via langchain-ollama
openai==1.99.9
    # via
    #   -r requirements.in
    #   langchain-openai
orjson==3.11.2
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.10.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
pandas==2.3.1
    # via -r requirements.in
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.7
    # via
    #   langchain
    #   langchain-core
    #   langgraph
    #   langsmith
    #   ollama
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via langchain-community
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2025.7.34
    # via tiktoken
requests==2.32.4
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.43
    # via
    #   langchain
    #   langchain-community
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
tiktoken==0.11.0
    # via langchain-openai
tqdm==4.67.1
    # via openai
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   anyio
    #   exceptiongroup
    #   langchain-core
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
xxhash==3.5.0
    # via langgraph
yarl==1.20.1
    # via aiohttp
zstandard==0.23.0
    # via langsmith
